Fix for out of bounds read in torch flatbuffer loader (#108433)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/108433

This diff fixes an out of bounds read issue that was introduced because we were accessing a flatbuffer object before verifying it.

This diff moves the accessing code after the verification call to remove the potential for the out of bounds read condition.

Test Plan: Out of bounds read crash no longer reproduces

Differential Revision: D48875827

fbshipit-source-id: 45cbf701102dc5af6f42f10698573b1056147bcb